Add tests & fixes for epp contact
