+++ app/models/domain.rb
@@ -1,5 +1,6 @@
 class Domain < ApplicationRecord
   include UserEvents
+  include Roids
   include Versions # version/domain_version.rb
   include Domain::Expirable
   include Domain::Activatable
@@ -32,8 +33,6 @@ class Domain < ApplicationRecord
   accepts_nested_attributes_for :admin_domain_contacts,  allow_destroy: true, reject_if: :admin_change_prohibited?
   has_many :tech_domain_contacts
   accepts_nested_attributes_for :tech_domain_contacts, allow_destroy: true, reject_if: :tech_change_prohibited?
-
-  ID_CHAR_LIMIT = 8
 
   def registrant_change_prohibited?
     statuses.include? DomainStatus::SERVER_REGISTRANT_CHANGE_PROHIBITED
@@ -330,15 +329,6 @@ class Domain < ApplicationRecord
       domain = self.find_by_name unicode
     end
     domain
-  end
-
-  def roid
-    id_size = id.to_s.size
-    if id_size <= ID_CHAR_LIMIT
-      "EIS-#{id}"
-    else
-      roid_with_prefix(id_size)
-    end
   end
 
   def puny_label
@@ -740,14 +730,5 @@ class Domain < ApplicationRecord
 
   def self.uses_zone?(zone)
     exists?(["name ILIKE ?", "%.#{zone.origin}"])
-  end
-
-  private
-
-  def roid_with_prefix(id_size)
-    id_delta = id_size - ID_CHAR_LIMIT
-    id_prefix = id.to_s.split(//).first(id_delta).join('').to_s
-    id_postfix = id.to_s.split(//).last(id_size - id_delta).join('').to_s
-    "EIS#{id_prefix}-#{id_postfix}"
   end
 end
